Fix tests, add utility methods, remove extra changes
