+++ indexing-service/src/main/java/org/apache/druid/indexing/common/task/NativeCompactionRunner.java
@@ -88,7 +88,7 @@ public class NativeCompactionRunner implements CompactionRunner
       CompactionTask compactionTask
   )
   {
-    return new CompactionConfigValidationResult(true, null);
+    return CompactionConfigValidationResult.success();
   }
 
   /**
